v4/asm--写入.data，无语法错误，无明显的数据错误

+++ v4/asm/common.h
@@ -109,7 +109,9 @@ typedef enum{
 }OFFSET_TYPE;
 
 typedef struct{
-	int value;
+	// int value;
+	// unsigned int value;
+	long long int value;
 	OFFSET_TYPE type;
 } NumericData;
